wellwork_20190911

+++ ros/occupancy/include/ProjectMap.h
@@ -56,7 +56,7 @@ namespace ProjectMap
         unsigned int all_pts_pub_gap_ = 0;
         unsigned int pub_count_ = 0;
 
-        cv::Mat orb2ros = (cv::Mat_<int>(3,3) << 0,0,1,-1,0,0,0,-1,0);
+        cv::Mat orb2ros = (cv::Mat_<float>(3,3) << 0,0,1,-1,0,0,0,-1,0);
     };
 
     class Map
